feat: se agrega alertas

+++ src/app/modules/auth/ingresar/ingresar.component.ts
@@ -1,45 +1,72 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../services/auth.service';
 import { Router } from '@angular/router';
-
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-ingresar',
   templateUrl: './ingresar.component.html',
-  styleUrls: ['./ingresar.component.sass']
+  styleUrls: ['./ingresar.component.sass'],
 })
 export class IngresarComponent implements OnInit {
+  cedula: number;
 
-  cedula:number 
-
-  constructor(
-    private authService: AuthService,
-    private router: Router
-  ) { }
+  constructor(private authService: AuthService, private router: Router) {}
 
   ngOnInit(): void {
-    sessionStorage.clear()
+    sessionStorage.clear();
   }
 
-  ingresar(){
+  ingresar() {
 
+    // Verificar si la cédula está vacía
+  if (!this.cedula) {
+    // Mostrar mensaje de error y evitar realizar la solicitud
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      showConfirmButton: false,
+      timer: 3000,
+      timerProgressBar: true,
+    });
 
-    this.authService.getUsuario(this.cedula).subscribe(
-      res => {
+    Toast.fire({
+      icon: 'error',
+      title: 'Por favor ingresa tu cédula',
+    });
 
-        if(Object.keys(res).length > 0 ){
+    return;
+  }
 
+  this.authService.getUsuario(this.cedula).subscribe(
+      (res) => {
+        if (Object.keys(res).length > 0) {
           const resString = JSON.stringify(res);
-          sessionStorage.setItem("session", resString)          
+          sessionStorage.setItem('session', resString);
           this.router.navigate(['servicios']);
-
         }
-      
-      },error => {
-        console.error(error);
-      }
-    )
-  }
+      },
+      (error) => {
+        // console.error("holaad");
 
+        this.router.navigate(['login']);
+
+        //mensaje
+
+        const Toast = Swal.mixin({
+          toast: true,
+          position: 'top-end',
+          showConfirmButton: false,
+          timer: 3000,
+          timerProgressBar: true,
+        });
+
+        Toast.fire({
+          icon: 'error',
+          title: 'No te encuentras registrado',
+          footer: 'Te invitamos a registrarte para ingresar',
+        });
+      }
+    );
+  }
 }
